requirements.txt
asgiref==3.7.2
Django==4.2.5
django-ckeditor==6.7.0
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
psycopg2==2.9.7
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
